--- Tortillapp-web/Pages/FirstView.cshtml.cs
@@ -1,33 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.EntityFrameworkCore;
-using Tortillapp_web.Data;
-using Tortillapp_web.Model;
-
-namespace Tortillapp_web.Pages
-{
-    public class FirstViewModel : PageModel
-    {
-        private readonly Tortillapp_web.Data.tortillaContext _context;
-
-        public FirstViewModel(Tortillapp_web.Data.tortillaContext context)
-        {
-            _context = context;
-        }
-
-        public IList<RecipeInfo> RecipeInfo { get; set; } = default!;
-
-        public async Task OnGetAsync()
-        {
-            if (_context.RecipeInfos != null)
-            {
-                RecipeInfo = await _context.RecipeInfos
-                .Include(r => r.User).ToListAsync();
-            }
-        }
-    }
-}
