+++ Tortillapp-web/Pages/Recipe/Index.cshtml
@@ -5,7 +5,6 @@
     ViewData["Title"] = "Index";
 }
 
-
 <div class="container">
     <div class="row mt-3">
         <div class="row text-center align-items-center justify-content-center">
@@ -57,8 +56,7 @@
                 </div>
                  <div class="row">
                     <div class="col">
-                    @Html.DisplayFor(modelItem => item.RecipeStatus)
-                    
+                        @Html.DisplayFor(modelItem => item.RecipeStatus)
                     </div>
                 </div>
             </div>
