+++ Tortillapp-web/Pages/Register.cshtml
@@ -10,8 +10,10 @@
         <div class="col-sm">
             <h1 class="text-center">Se parte de la comunidad</h1>
             <div class="row h-75 aling-items-center justify-content-center">
-                <div class="col-auto">
-                    <label for="backto" class="form-label">¿Ya eres parte? <a href="/Login">Inicia sesión</a></label>
+                <div class="col div_base">
+                    <div class="div_translate">
+                        <label for="backto" class="form-label">¿Ya eres parte? <a href="/Login">Inicia sesión</a></label>
+                    </div>
                 </div>
             </div>
         </div>
