+++ Tortillapp-web/Pages/Shared/_Layout.cshtml
@@ -97,7 +97,7 @@ else { isLogged = 0; }
             </nav>
         </div>
     </header>
-    <div class="container-fluid">
+    <div class="container-fluid no-gutters">
         <main role="main" class="pb-3">
             @RenderBody()
         </main>
