+++ Tortillapp-web/Pages/_IndexTags.cshtml
@@ -1,0 +1,8 @@
+﻿@model IndexModel
+
+@foreach (var tag in Model.Tags)
+{
+    <label>
+        <a class="mb-2 tag_base tag_deco button_deco">@Html.DisplayFor(modelItem => tag.TagName)</a>
+    </label>
+}
